Add checks to Settings cog
