+++ tdx/settings.py
@@ -3,7 +3,7 @@ import logging
 from typing import Dict, List
 
 import discord
-from redbot.core import commands, Config
+from redbot.core import checks, commands, Config
 from redbot.core.bot import Red
 from redbot.core.i18n import Translator
 from redbot.core.utils import chat_formatting as cf
@@ -18,11 +18,77 @@ class Settings(commands.Cog):
         self.bot: Red = bot
         self.config: Config = config
 
+    @commands.command(name="quickstart")
+    @checks.mod_or_permissions(manage_guild=True)
+    @checks.bot_in_a_guild()
+    async def quickstart(self, ctx: commands.Context) -> None:
+        await ctx.send(_("Looking for team roles..."))
+
+        try:
+            mystic_role: discord.Role = min(
+                [x for x in ctx.guild.roles if _("Mystic").casefold() in x.name.casefold()]
+            )
+        except ValueError:
+            mystic_role = None
+        if mystic_role:
+            await getattr(self.config.guild(ctx.guild), "mystic_role").set(mystic_role.id)
+            await ctx.send(
+                _("`{key}` set to {value}").format(key="mystic_role", value=mystic_role.mention)
+            )
+
+        try:
+            valor_role: discord.Role = min(
+                [x for x in ctx.guild.roles if _("Valor").casefold() in x.name.casefold()]
+            )
+        except ValueError:
+            valor_role = None
+        if valor_role:
+            await getattr(self.config.guild(ctx.guild), "valor_role").set(valor_role.id)
+            await ctx.send(
+                _("`{key}` set to {value}").format(key="valor_role", value=valor_role.mention)
+            )
+
+        try:
+            instinct_role: discord.Role = min(
+                [x for x in ctx.guild.roles if _("Instinct").casefold() in x.name.casefold()]
+            )
+        except ValueError:
+            instinct_role = None
+        if instinct_role:
+            await getattr(self.config.guild(ctx.guild), "instinct_role").set(instinct_role.id)
+            await ctx.send(
+                _("`{key}` set to {value}").format(
+                    key="instinct_role", value=instinct_role.mention
+                )
+            )
+
+        await ctx.send(_("Looking for TL40 role..."))
+
+        try:
+            tl40_role: discord.Role = min(
+                [x for x in ctx.guild.roles if _("Level 40").casefold() in x.name.casefold()]
+            )
+        except ValueError:
+            tl40_role = None
+        if tl40_role:
+            await getattr(self.config.guild(ctx.guild), "tl40_role").set(tl40_role.id)
+            await ctx.send(
+                _("`{key}` set to {value}").format(key="tl40_role", value=tl40_role.mention)
+            )
+
+        await ctx.send(_("That's it for now."))
+
+        settings: Dict = await self.config.guild(ctx.guild).all()
+        settings: str = json.dumps(settings, indent=2, ensure_ascii=False)
+        await ctx.send(cf.box(settings, "json"))
+
     @commands.group(name="tdxset", aliases=["config"])
     async def settings(self, ctx: commands.Context) -> None:
         pass
 
     @settings.group(name="guild", aliases=["server"])
+    @checks.mod_or_permissions(manage_guild=True)
+    @checks.bot_in_a_guild()
     async def settings__guild(self, ctx: commands.Context) -> None:
         if ctx.invoked_subcommand is None:
             settings: Dict = await self.config.guild(ctx.guild).all()
@@ -208,6 +274,8 @@ class Settings(commands.Cog):
             )
 
     @settings.group(name="channel")
+    @checks.mod_or_permissions(manage_guild=True)
+    @checks.bot_in_a_guild()
     async def settings__channel(self, ctx: commands.Context) -> None:
         if ctx.invoked_subcommand is None:
             settings: Dict = await self.config.channel(ctx.channel).all()
@@ -235,3 +303,16 @@ class Settings(commands.Cog):
                     key=f"channel[{ctx.channel.id}].profile_ocr", value=value
                 )
             )
+
+    @settings.group(name="user", aliases=["member"])
+    async def settings__user(self, ctx: commands.Context) -> None:
+        if ctx.invoked_subcommand is None:
+            if ctx.guild:
+                settings: Dict = {
+                    **await self.config.member(ctx.author).all(),
+                    **await self.config.user(ctx.author).all(),
+                }
+            else:
+                settings: Dict = await self.config.user(ctx.author).all()
+            settings: str = json.dumps(settings, indent=2, ensure_ascii=False)
+            await ctx.send(cf.box(settings, "json"))
